# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VBALANCE Documentation",
        "description": "Documentation for VBALANCE",
        "version": "2"
    },
    "servers": [
        {
            "url": "/api/v2",
            "description": "Relative V2"
        },
        {
            "url": "/api/v1",
            "description": "Relative V1"
        }
    ],
    "paths": {
...
        "/ia-system/history": {
            "get": {
                "tags": [
                    "IA-System"
                ],
                "description": "История депозитов ia-system. Статусы: 1 - ожидание, 2 - активен, 3 - закончился, 4 - выведен.",
                "operationId": "70369811fbdea87427294fd52d63021b",
                "parameters": [
                    {
                        "name": "field",
                        "in": "query",
                        "example": "created_at,status,count_months,amount,percent"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "example": "asc,desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/ia-system/reopen": {
            "put": {
                "tags": [
                    "IA-System"
                ],
                "description": "Продлить депозит ia-system",
                "operationId": "ba5416f0e4224e23eac0f80c6719cd75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/ia-system/change": {
            "put": {
                "tags": [
                    "IA-System"
                ],
                "description": "Изменить депозит ia-system",
                "operationId": "ac3bc66bfeed70624d01d93e115a2f76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "example": "123"
                    },
                    {
                        "name": "count_months",
                        "in": "query",
                        "example": "6|12|18"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/user/{user}/statistic/users-list": {
            "get": {
                "tags": [
                    "User.Statistic"
                ],
                "description": "Список пользователей. Кэш обновляется на offset=0",
                "operationId": "UserStatisticUserList",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение (Кол-во загруженных записей (next_offset))"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Лимит"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки (line,status,first_line_count,organization_count,organization_total_sail_sum,organization_total_sail_sum)"
                    },
                    {
                        "name": "sort_reverse",
                        "in": "query",
                        "description": "Обратная сортировка (DECS) true/false"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserStatisticUsersListResource"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "all_count": {
                                                    "type": "integer"
                                                },
                                                "count": {
                                                    "type": "integer"
                                                },
                                                "next_offset": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserStatisticUsersListResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "login": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "line": {
                        "description": "Номер линии",
                        "type": "integer"
                    },
                    "first_line_count": {
                        "description": "Размер первой линии",
                        "type": "integer"
                    },
                    "organization_count": {
                        "description": "Размер организации",
                        "type": "integer"
                    },
                    "organization_total_sail_sum": {
                        "description": "Товарооборот организации",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}